# Optimize applyToSelectedNoThrow