Summary:
ServiceLab (Meta internal perf testing tool) catched a regression in  https://github.com/facebookincubator/velox/pull/3729
, investigation shows its related to applyToSelectedNoThrow

this diff fix it.
before:
```
============================================================================
[...]arks/basic/LikeFunctionsBenchmark.cpp     relative  time/iter   iters/s
============================================================================
wildcardExactlyN                                            6.10ms    163.93
wildcardAtLeastN                                            5.94ms    168.31
suffixPattern                                              12.73ms     78.57
tpchQuery16Part                                            13.13ms     76.18
```

after:
```
============================================================================
[...]arks/basic/LikeFunctionsBenchmark.cpp     relative  time/iter   iters/s
============================================================================
wildcardExactlyN                                            5.30ms    188.64
wildcardAtLeastN                                            5.10ms    196.24
suffixPattern                                              10.83ms     92.30
tpchQuery16Part                                            10.82ms     92.39
```

Reviewed By: mbasmanova

Differential Revision: D42711902